Add a KG-LLM Complete Paper (EMNLP 2023 Findings)
We add our work KICGPT(Knowledge In Context
with GPT):
The KICGPT framework integrates conventional Knowledge Graph Completion (KGC) methods with the capabilities of Large Language Models (LLMs) to improve KGC, particularly in the area of link prediction. Initially, it utilizes standard KGC techniques to generate initial predictions, which are then refined using LLMs that re-assess and prioritize the results. This refinement process leverages In-Context Learning (ICL), where knowledge graph information is strategically embedded into the prompt. The KICGPT framework is notably proficient in mitigating the challenges associated with the long-tail distribution of entities and relations within knowledge graphs.